Combined middleware and errorMiddleware files
