+++ src/core/lib/decorators/index.ts
@@ -1,5 +1,4 @@
 export * from './class';
-export * from './errorMiddleware';
 export * from './method';
 export * from './middleware';
 export * from './wrapper';
